Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

When you touch an album from the "Albums" tab then press the "up" button in the ActionBar, the user should be taken to the corresponding artist page. You removed the few lines that took care of this in "TracksBrowser".

You're right about touching an artist, then album, then the "back" button though.

Patch-set: 1