Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Regarding to theming widget, it seems impossible to me looking that there's not even an import on AppWidget41 or AppWidget42 for utils.ThemeUtils. As a workaround "overflowLight(Context mContext)" could be used in those classes for overriding current icons where they are initially set. Or better, add theming support to widgets. But I leave that to you :)

Patch-set: 2