Fix Issue 5547: Apollo will not play MP3 files when using Root Explorer
Apollo doesn't have an activity registered for support "file" schemas, that are used by file
managers when using the method setDataAndType of Intent class.

This changes adds a new activity (PlayExternal) that registers this kind of schema, and pass
the request to AudioPlayerHolder. Allow direct play the file, or enqueue in the current track list

Change-Id: I25b728700bd0d2d58d31e4de4bed81ee3e368e0a