Added lyrics support
This change adds a feature to read a song's lyrics in the player activity while
playing a song.

It works both online and offline.
It will use the jaudiotagger library to read the song from the file's metadata.
If they are not found it will use an online provider to search for the lyrics
and show them. (Currently there is only one provider, more providers can easely
be added with the LyricsProviderFactory and LyricsProvider Interface)

ApolloService and MusicUtils have been modified to provide the current song's
file path. Write permission has been added to save the downloaded lyrics into
the metadata for offline use.

To show/hide the lyrics just tap on the album art in the player.
A screenshot is available here: http://kcy.me/a1ek
And a test apk is here: http://kcy.me/a1em

This is my first time with gerrit, I hope I did everything all right.