Update patch set 2
Patch Set 2:

This is really nice. This is the first major contribution to Apollo since we first merged it with CM. I really appreciate it. I'd like to make it optional, so I'm going to hold off for a little while for that reason, and because I'm preparing to publish quite a few new changes very soon. This is really exciting though, I used your test app briefly and everything looks and works smoothly.

Patch-set: 2