Update patch set 2
Patch Set 2: (4 inline comments)

See inline comments on each file

Also, a lot of your code is using tabs when it should be 4-spaces. You should be able to configure this in your editor. If you view each file in the review the parts where there are red double arrows are the parts that are using tabs.

Patch-set: 2