Update patch set 1
Patch Set 1: Do not submit

I've played around with the new notifications quite a bit and although I'd like to use them, I don't think we should, due to the lack of flag or method to indicate that you don't want the notification expanded by default. So, if you're using the new Big Picture Style and open the notification bar, Apollo's notification will be expanded by default, this is up to the System UI. If I could control them completely, I would have them show music controls while collapsed and extra controls, such as "share" and "favorite", while expanded with the artist picture. I suppose I could write some patches to the System UI and Notifications classes to allow a flag to be set to indicate when they should be expanded by default, but I don't have much time to work on that right now. So, until that flag can be set, it seems to me that the new notifications are just going to feel intrusive to the users.

Patch-set: 1