Update patch set 1
Patch Set 1:

The expanded layout of the notification is intended to display "more" information. In Apollo's case, that would be a picture of the currently playing artist. If you don't add the extra actions, then the user would need to collapse the notification every time the song changes to control their music from the notification. It then becomes obvious that you would need to add the action buttons to adjust for this. So, you add the next, play, and collapse buttons to the expanded layout, but what do you show when it's collapsed now? Do you show the same buttons? If you set the notification priority to the default, then other notifications will collapse Apollo's and in that case showing the next, play, and collapse buttons in the collapsed layout is necessary, otherwise users will have to then expand it to control their music every time the song changes. You would have to set Apollo's notification priority to the highest flag to keep this from happening and the music notification is not important enough to be set that high, especially to accommodate for something like a picture of the artist. Also, the next, play, and collapse buttons aren't "more" information. They belong in the collapsed layout and actions like "share", "favorite", "add to playlist", or "shop" belong in the expanded layout. In that case, the user would then have to collapse the notification every time the song changes to control their music. 

So, that's how I see it and why I think they're inconvenient and should be added later when the notifications can be finely tuned. I think they're cool too and should be added eventually.

Patch-set: 1