Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

It's called Play Store in most of the other countries, so let us also use that. Dunno why the Nordic countries insists on calling it Butikk, as it just sounds silly in my ears. It's like calling PayPal for BetalingsVenn or something. :)

Patch-set: 1