Update patch set 1
Patch Set 1:

There is a number of things in Apollo which I'd like to improve but I'd like to get this rewrite out of the way first. For example I'd like to modify the album image displaying logic to use the cover file provided with the album if available. With the current image downloading logic scattered all around the application it'd require a lot of repetitive code.

Patch-set: 1