Update patch set 1
Patch Set 1:

Andrew, It's a shame I haven't seen your changes before I started working on this. Could you push your changes (even if that's work in progress) to some repo/branch so that I can have a look and contribute? As I wrote in my previous comment there is a number of features I'd like to add. To do that I'll need a visibility of the latest development code.

Patch-set: 1