Update patch set 1
Patch Set 1:

How is it going with your rewrite?
Last time we spoke you told me that your hard deadline for finishing this change is 6th Oct. It's 30th now and I can't see any new code. Apollo is a promising application with the nice UX but it's still quite immature/unpolished. IMHO to move it forward you should try and build a community of developers contributing to this project. Freezing the code for two months surely won't help.

Patch-set: 1