Update patch set 1
Patch Set 1:

I've been working for at least twelve hours everyday on this for months, so any rewrite started then not now, that's no exaggeration. The plan is to work in the open, but before I can do this I have to have fully functional app. I can't push an unfinished code. Look, I'm sorry about the time it's taking me to do this. I rushed last time to get things done and I this time I decided to take my time. I'm trying to write code as professionally as possible so when I do push the changes everyone who is interested in helping can immediately. I realize the code right now is "immature" and not very professional, but I can definitely assure you I'm doing the best I can. I'll push the changes before the week is up. This isn't going to be the situation all the time.

Patch-set: 1