Update patch set 1
Patch Set 1:

I appreciate your hard work and I hope the result will be great. What you are saying sounds scary to me though!

You spent more than 1000hrs and haven't released anything to the public? I guess you don't employ a dozen of full time testers with tens of different devices, do you? If not it means that you are going to release a huge chunk of UNTESTED CODE. This can only lead to trouble.

I disagree that you "can't push an unfinished code". The code is never finished, there is always something to improve or to fix. Have a look at Agile software development methodologies.

If the release is scheduled next week I think we should do one of the following:
1. merge your stuff NOW so that community can start testing, fixing prioritising
2. put your changes on hold for now, concentrate of small and safe fixes of current code for this release and then include your rewritten code in next release.

I haven't seen your new code so I can't tell which option is better but for me delaying the changes till the last day and then releasing an untested code is not an option.

Patch-set: 1