Update patch set 1
Patch Set 1:

It's not like that. I'm not releasing untested code, I test on four different devices and three versions of Android. I'm finishing, putting it all together. You're right, code is never finished. It can always be improved on in some way or another, but there's a difference between that unfinished code and unfinished code that can't be used. I'm not going to push changes and cross my fingers, I'm going to push them in confidence. Give me a little slack, this will be done and things will be good.

Patch-set: 1