Update patch set 1
Patch Set 1:

Andrew, are you going to publish your changes now when you released Apollo+? What are your plans for further development of Apollo?

Evan, I'm glad you find this patch useful. As I mentioned before, it's a preparation&cleanup patch. There is a number of relatively simple fixes/improvements which can be added on top of this patch.

Patch-set: 1