Update patch set 1
Patch Set 1:

I'm not going to abandon the CM version. Most of the changes I push to the Play Store will be pushed here as well. I'm not dropping the code and I plan on working in the open from here on out. Like I've been telling you man, I just needed to get things to where they are now. The development for Apollo is really about to start. It's at a solid, stable point now and I'm ready to see contribution and the app flourish. This was my intention all along.

Patch-set: 1