Update patch set 1
Patch Set 1:

Whole Action Bar Sherlock library is missing. You'll need to add the compiled *.jar to the lib folder or copy the sources.

The patch is quite overwhelming. I'm trying to figure out what made the code size grew from 16k LOC to 36k LOC. It'd be easier to navigate if you could keep the external code (components, libraries, etc.) in a separate packages, preferably original authors packages so that it's easy to track it back.

Patch-set: 1