Update patch set 1
Patch Set 1:

Andrew,

Have you removed Apollo(+) from the Play Store? Why?

I'm on holiday next week but when I'm back I'll be tempted to fork the current CM10 version of Apollo. It's a music player I use every day and there are many small but annoying bugs. IMO this project needs loads of small bugfixes not another rewrite.

I haven't got a chance to look at your new version submitted to master (it doesn't build) but for me it's hard to justify maintaining 30k lines of code when the functionality is almost the same as the previous version with 15k LOC. Could you explain this code bloat?

Patch-set: 1