Update patch set 1
Patch Set 1:

Thanks for rebasing my patch. Before it gets merged someone (Steve?) has to decide what to do with Andrew's changes in master. IMO they are not ready to be merged to production branches. On the other hand it'll get even harder to merge if we continue development on the current codebase.

Personally I'd prefer for Andrew's changes not to be merged. Instead we could manually copy the new features taking care for the code not to bloat (the code in master is two times bigger than ICS). Accepting these changes in smaller chunks would also give us an opportunity to revert some changes if they are not welcome by the users. e.g. Users complained about new layout in Play Store comments.

Steve, given lack of Andrew's activity who can take over merging patches to Apollo?

Patch-set: 1