Update patch set 1
Patch Set 1:

Hah, you are right on that one, I wasn't paying attention to that small detail.
I'm not comfortable with sending 2 intents there - should I simply add a NOW_IN_FOREGROUND extra to the intents and modify MusicPlaybackService.onStartCommand or onHandleCommand to check that extra? Or is there a better solution?

Patch-set: 1