Update patch set 6
Patch Set 6: Looks good to me, but someone else must approve

That's an excuse, right? ;)
Anyways, you still have a call to buildNotification() or killNotification() that's not needed (you press the pause button, you'll have a call to buildNotification() and in the handleCommandIntent the notification gets killed later) there but anyways the patch seems fine.

Patch-set: 6