Update patch set 1
Patch Set 1: (2 inline comments)

Most of the translations appear to be correct. However I commented on two, which I think fits better. One place you use 'albummer' (which is the correct usage as far as I know) and on line 58 you use 'albums'.
Line 111 is a bit misleading 'Bannerdesign for ikon' is not what I would translate it to. It is not just 'Bannerdesign' but also 'Ikondesign'.
Apart from these two, the rest seems to be fine.

Patch-set: 1