Update patch set 1
Patch Set 1:

Unfortunately, the drawables alone don't help for the large and very large font settings. Text is still cut off with the current layouts in that case. I'll test the very long text case today.

Is there any specific reason for the -2 instead of -1, given that you say it's a good idea?

Patch-set: 1