Update patch set 1
Patch Set 1:

The only reason I had this to begin with is because the RCC overrides all other lockscreen widgets, which can get kind of annoying if you never use them or want to see a different widget when you turn your screen on. I suppose a majority of users take advantage with the playback controls though. Considering that, if you still feel like that option should be tossed I don't really have a problem with it.

Patch-set: 1