Apollo: Make albums with the same name to use respective album-art/date/etc.
It may happen that two (or more) albums have the same name but each of them
comes from different artist. This case was not handled correctly causing with
the same album art to be used for all albums.

This patch makes sure that the proper album data is fetched for album. It's
achieved by either adding album artist (i.e., for cached images) or using
album id (unique) where possible (not causing too much code churn).

Change-Id: I181a28d80a18207101c613cca390f18d88dd556a