Apollo: Added triple click prev track
Added a triple click for the headset button so that Apollo goes back to
the previous track.

The current implementation means that the triple click will first be
preceded by a double click, which will jump to the next track. If a
triple click is detected, it will jump back 2 tracks, which is
effectively the one before. However, the user might hear a split second
of audio from the next track.

Change-Id: I46b14136ec471673a4a139dc2f4b50ecacaccc1e