Update patch set 2
Patch Set 2: Code-Review-1

I don't like this implementation; it feels rather hacky. If we want triple click functionality, we should implement it properly: Defer the single/double click action until the total number of clicks is known. Something like the following:

- introduce a new message (HEADSET_CLICK) for the handler
- when a click is detected, passing in the total number of clicks detected so far
- if number of clicks < 3, use DOUBLE_CLICK_TIMEOUT as delay, otherwise use no delay
- when processing the message, the handler can use the total number of clicks to determine which command to send

This will introduce a 300ms delay to the processing of the single/double clicks of the headset button, but I don't think that's a problem in practise.

Patch-set: 2