Update patch set 2
Patch Set 2:

The delay is normal (it's the wait time for the next click). For the playback problem, I think there is a wakelock problem somewhere in the audio stack. The second 'next' attempt looks like the CPU goes to sleep in the middle of changing tracks. Does it work if you add a 'return;' at the beginning of releaseWakeLockIfHandlerIdle? If yes, that would confirm my suspicion.

I'm not sure why you have the problem and I don't, but it's likely that it's because you are using tunnel audio and I am not (Nexus 4), so my phone's CPU is running anyway.

Can you please add a log output at the end of MusicPlaybackService.onStartCommand() and see whether that's reached in a timely manner? If no, that points towards one needing to hold a wake lock while starting the playback service from the receiver. If yes, something in the audio stack isn't holding the CPU awake for long enough.

Patch-set: 2