Update patch set 4
Patch Set 4:

I vectorized the submitted drawable and made an xxhdpi version. I also included the PSD with the vector version as a Smart Object for easy scaling as well as the ai file if needed later.

https://www.dropbox.com/s/rtqyubqt1eikn78/default_artwork.zip

Please push the xxhdpi png to this change. I didn't feel like doing it this early in the morning.

Patch-set: 4