Update patch set 1
Patch Set 1: Code-Review-1

@Danny Baumann:
I've fixed the line wrapping issue and I'll post a newer version a bit later. I've also dropped the different theme key, although I kinda' liked having that label in a different color :).

Some screenshots:
(before)
http://i.imgur.com/kLtHxrY.png

(after)
http://i.imgur.com/WGlLegP.png
http://i.imgur.com/m4RPleG.png

@Jacob Müller:
I've fixed the nits -- I was trying to follow the style of the related code in the application.


I'll post a version that, in addition to addressing the review comments, adds that label in more views in Apollo. I won't be adding it to the search results because the duration doesn't seem to be readily available.

Patch-set: 1