Update patch set 1
Patch Set 1:

Well my car is actually a japanese import and the headunit only goes to 90mhz which makes it mostly useless for New Zealand radio stations. So I have a Bluetooth FM transmitter, which is what my phone is connecting to.

But even in my old car where the headunit had proper bluetooth built in, my phone never fired up the media player and started playing when it connected

So you say this is the incorrect approach because it should be handled at the system level and not a media player level? If so, I have never been able to get this to work at the system level

I guess the whitespace issues are due to the default Eclipse settings, I can see them clearly now when looking at the diffs

Patch-set: 1