Update patch set 2
Patch Set 2:

That's what I meant by 'fix mediascanner' :) Using the folder as condition for determining album relationships is not a good idea, if it's really doing that. The media provider interface definitely doesn't mandate that, so if we fix this in mediascanner, all audio player apps profit from it and the code in Apollo becomes easier.

Patch-set: 2