Update patch set 3
Patch Set 3:

The async load of the artwork makes not sense to me, because to know if the mp3 has artwork one has to read the id3tags, which give a byte[] (the artwork itself), so the artwork is loaded. The load of the image to bitmap is currently asynchronous.

The cache makes sense to me as well. I will implement it.

Patch-set: 3
Reviewer: Jorge Ruesga <3134@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0